backend/requirements.txt
fastapi>=0.110.0,<0.120.0
uvicorn>=0.27.0,<0.30.0
pydantic>=2.6.0,<3.0.0
email-validator>=2.1.0,<3.0.0
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0
bcrypt>=4.0.1,<5.0.0
PyJWT>=2.8.0,<3.0.0
typing-extensions>=4.9.0,<5.0.0
mysqlclient>=2.2.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pymysql>=1.1.0,<2.0.0
